+++ tests/testthat/test_filter_samples.R
@@ -1,10 +1,5 @@
 test_that("Sample filtering tests", {
-    hVcfFileDir <- system.file("extdata", package = "rPHG2")
-    hVcfFiles   <- list.files(hVcfFileDir, pattern = ".h.vcf$", full.names = TRUE)
-    locCon      <- PHGLocalCon(hVcfFiles)
-    graph       <- buildHaplotypeGraph(locCon)
-    pds         <- readPhgDataSet(graph)
-
+    pds <- makeExamplePds()
 
     obsFiltPds01 <- filterSamples(pds, "LineB")
     obsFiltPds02 <- filterSamples(pds, c("LineA", "LineX", "LineB"))