+++ tests/testthat/test_read_samples.R
@@ -27,11 +27,7 @@ test_that("Sample reading tests from JVM object", {
 
 
 test_that("Sample reading tests from PHGDataSet", {
-    hVcfFileDir <- system.file("extdata", package = "rPHG2")
-    hVcfFiles   <- list.files(hVcfFileDir, pattern = ".h.vcf$", full.names = TRUE)
-    locCon      <- PHGLocalCon(hVcfFiles)
-    graph       <- buildHaplotypeGraph(locCon)
-    pds         <- readPhgDataSet(graph)
+    pds <- makeExamplePds()
 
     obsSamples <- readSamples(pds)